+++ vcd/resource_vcd_edgegateway_vpn_test.go
@@ -5,8 +5,8 @@ package vcd
 import (
 	"testing"
 
-	"github.com/hashicorp/terraform/helper/resource"
-	"github.com/hashicorp/terraform/terraform"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
 
 func TestAccVcdVpn_Basic(t *testing.T) {